French comment

+++ app/controllers/addressController.js
@@ -2,13 +2,11 @@ const { Article, Category, Size, User, Order, Address, ArticleHasSize } = requir
 const sequelize = require('../database');
 
 const addressController = {
-    // renvoi toutes les address avec chacune leur user
+    // renvoi toutes les address avec chacun de leur user
     getAll: async (req, res) => {
         const { limit } = req.query;
         const addresses = await Address.findAll({
-            // attributes: {
-            //     exclude: ['created_at', 'updated_at']
-            // },
+
             include: [
                 {
                     association: 'address_user',
@@ -18,10 +16,10 @@ const addressController = {
         res.json(addresses);
     },
 
-    // renvoi l'address avec son user par rapport au user_id
+    // renvoi une address et son user
     getOne: async (req, res) => {
-        const user_id = req.params.id;
-        const address = await Address.findByPk(user_id, {
+        const { id } = req.params;
+        const address = await Address.findByPk(id, {
             include: [
                 {
                     association: 'address_user',
@@ -35,7 +33,7 @@ const addressController = {
         res.json(address);
     },
 
-
+    //cette fonction créée une addresse en fonction de son user id et la retourne 
     create: async (req, res) => {
 
         const newAddressData = {
@@ -49,30 +47,19 @@ const addressController = {
             lastname_address: req.body.lastNameAddress,
             user_id: req.body.userId
         };
-        console.log('newaddressdata :', newAddressData)
         const newAddress = await Address.create(newAddressData);
-        console.log('newaddress :', newAddress)
-
         const theAddressUser = await Address.findOne({
             where: { id: newAddress.id },
             attributes: {
                 exclude: ['created_at']
-            },
-            include: [{
-                association: 'address_orders',
-                include: [{
-                    association: 'orderArticles',
-                    order: [
-                        ['updated_at', 'ASC']
-                    ]
-                }]
-            }]
-
+            }
         });
 
         res.json(theAddressUser);
 
     },
+
+    // modification d'une address en fonction de son id
     updateById: async (req, res) => {
         const { id } = req.params;
         const newAddressData = {
@@ -87,24 +74,31 @@ const addressController = {
             user_id: req.body.userId
         };
         console.log('data renvoyée:', newAddressData)
-        //const oldUser = await Address.findOne({where:{id}});
+
         await Address.update({ ...newAddressData }, { where: { id } })
-        const newUser = await Address.findByPk(id)
-        res.json(newUser)
+        const newAddress = await Address.findByPk(id)
+        res.json(newAddress)
     },
+
+    // Suppression d'une addresse en fonction de son id
     deleteById: async (req, res) => {
-        try {
+       
             const { id } = req.params;
-            console.log('id:', id)
+            const address = await Address.findByPk(id);
+            if (!address) {
+                res.status(400).json(`l'addresse avec l'id ${id} n'existe pas et ne peut donc pas être supprimé`);
+                return next();
+            }
+
             const user = await Address.findByPk(id);
             Address.destroy({ where: { id } })
-            //await user.destroy();
-            res.json(`l'addresse avec l'id ${id} est bien supprimé`)
+
+            const addressExist = await Address.findByPk(id);
+            if (addressExist) {
+                res.status(400).json(`l'addresse avec l'id ${id} n'as pas était supprimé`);
+            };
+            res.json(`l'addresse avec l'id ${id} est bien supprimé`);
         }
-        catch {
-            res.status(400).json(`l'addresse avec l'id ${id} n'a pas pu être supprimé ou n'existe pas`)
-        }
-    }
 
 }
 module.exports = addressController;
